requirements.txt
# PSV to SQL Server Loader - Core Dependencies
# Essential packages required for the application to run

# Database connectivity and ORM
sqlalchemy>=1.4.0,<3.0.0

# SQL Server drivers
pyodbc>=4.0.0,<6.0.0
pymssql>=2.2.0,<3.0.0

# Data processing and manipulation
pandas>=1.3.0,<3.0.0

# Configuration file parsing
pyyaml>=5.4.0,<7.0.0